feature: create fetchUserInfo API
